requirements.txt
numpy
opencv-python
pillow
pywin32